pyproject.toml
[project]
name = "intelmarket"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "numpy>=2.3.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "scipy>=1.16.2",
    "streamlit>=1.50.0",
    "tabulate>=0.9.0",
]
